Split open_spiel_env tests: move game-specific cases to per-game env_test.py
Game-specific tests now live alongside each game at
tests/envs/open_spiel_env/games/<name>/env_test.py, matching the existing
coin_game_arena layout. The top-level test_open_spiel_env.py keeps only
framework-level coverage (env registration, strict mode, agent error,
game-parameter overrides, includeLegalActions, simultaneous dispatch).
Updates the onboard-open-spiel-game skill to point new games at the
per-game test file.
