+++ tests/envs/open_spiel_env/test_open_spiel_env.py
@@ -1,20 +1,17 @@
-import json
-import pathlib
-import random
+"""Framework-level tests for the open_spiel_env wrapper.
 
-import pokerkit  # noqa: F401
+Per-game tests live alongside each game in
+``tests/envs/open_spiel_env/games/<name>/env_test.py``. This file covers
+behavior of the open_spiel_env interpreter itself: env registration, strict
+mode, agent error / invalid action handling, game-parameter configuration,
+``includeLegalActions``, simultaneous-game dispatch, and serialized state.
+"""
+
 import pyspiel
 from absl.testing import absltest
-from open_spiel.python.games import pokerkit_wrapper  # noqa: F401
 
 from kaggle_environments import make
 from kaggle_environments.envs.open_spiel_env import open_spiel_env
-
-TEST_REPEATED_POKER_GAME_STRING = open_spiel_env.DEFAULT_REPEATED_POKER_GAME_STRING.replace(
-    "calcOddsNumSims=1000000",
-    "calcOddsNumSims=1",
-)
-
 
 # Expected that not all pyspiel registered games can be registered as Kaggle
 # envs, but should register at least this many
@@ -25,143 +22,6 @@ class OpenSpielEnvTest(absltest.TestCase):
     def test_envs_load(self):
         envs = open_spiel_env._register_game_envs([game_type.short_name for game_type in pyspiel.registered_games()])
         self.assertTrue(len(envs) > _REGISTERED_GAMES_THRESHOLD)
-
-    def test_coin_game_agent_playthrough(self):
-        env = make(
-            "open_spiel_coin_game",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_coin_game")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-
-    def test_coin_game_observation_is_json(self):
-        env = make(
-            "open_spiel_coin_game",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs_p0["phase"], "play")
-        self.assertEqual(obs_p0["current_player"], 0)
-        self.assertEqual(obs_p0["num_rows"], 8)
-        self.assertEqual(obs_p0["num_columns"], 8)
-        self.assertEqual(obs_p0["episode_length"], 20)
-        self.assertEqual(obs_p0["coin_colors"], ["a", "b", "c"])
-        self.assertEqual(len(obs_p0["board"]), 8)
-        self.assertTrue(all(len(row) == 8 for row in obs_p0["board"]))
-        # Player 0 sees their own preference but not opponent's.
-        self.assertIn(obs_p0["your_preference"], ["a", "b", "c"])
-        self.assertEqual(obs_p0["your_player_id"], 0)
-        self.assertNotIn("preferences", obs_p0)
-        # Both players are placed on the board.
-        self.assertIsNotNone(obs_p0["player_positions"]["0"])
-        self.assertIsNotNone(obs_p0["player_positions"]["1"])
-        self.assertFalse(obs_p0["is_terminal"])
-        self.assertEqual(obs_p0["move_number"], 0)
-
-    def test_coin_game_observation_hides_opponent_preference(self):
-        env = make("open_spiel_coin_game", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        obs_p1 = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(obs_p0["your_player_id"], 0)
-        self.assertEqual(obs_p1["your_player_id"], 1)
-        # Each player only learns its own preference; never the opponent's.
-        self.assertNotIn("preferences", obs_p0)
-        self.assertNotIn("preferences", obs_p1)
-
-    def test_coin_game_terminal_reveals_preferences_and_returns(self):
-        env = make(
-            "open_spiel_coin_game",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        final_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertEqual(final_obs["move_number"], 20)
-        self.assertEqual(final_obs["moves_remaining"], 0)
-        self.assertIn(final_obs["winner"], [0, 1, "draw"])
-        self.assertEqual(len(final_obs["returns"]), 2)
-        # On terminal, both preferences are revealed for scoring inspection.
-        self.assertIn("0", final_obs["preferences"])
-        self.assertIn("1", final_obs["preferences"])
-
-    def test_coin_game_invalid_action(self):
-        env = make("open_spiel_coin_game", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_dark_hex_agent_playthrough(self):
-        env = make(
-            "open_spiel_dark_hex",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_dark_hex")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        rewards = playthrough["rewards"]
-        self.assertEqual(sorted(rewards), [-1.0, 1.0])
-
-    def test_dark_hex_manual_playthrough(self):
-        env = make("open_spiel_dark_hex", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 1}, {"submission": -1}])  # p0: b1
-        env.step([{"submission": -1}, {"submission": 0}])  # p1: a1
-        env.step([{"submission": 4}, {"submission": -1}])  # p0: b2
-        env.step([{"submission": -1}, {"submission": 3}])  # p1: a2
-        env.step([{"submission": 7}, {"submission": -1}])  # p0: b3 (wins)
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [1, -1])
-        final_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertEqual(final_obs["winner"], "x")
-
-    def test_dark_hex_observation_hides_opponent(self):
-        env = make("open_spiel_dark_hex", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 4}, {"submission": -1}])  # p0: b2 (center)
-        # After p0's move, p1's view should not see the new x piece.
-        obs_p1 = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(obs_p1["board"], [["."] * 3] * 3)
-        # p0's view should see their own x piece.
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs_p0["board"][1][1], "x")
-
-    def test_dark_hex_invalid_action(self):
-        env = make("open_spiel_dark_hex", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
 
     def test_strict_mode_agent_error_keeps_per_player_status(self):
         env = make(
@@ -218,407 +78,6 @@ class OpenSpielEnvTest(absltest.TestCase):
         self.assertIn("actionString", played_action)
         self.assertTrue(played_action["actionString"])
 
-    def test_gin_rummy_agent_playthrough(self):
-        env = make(
-            "open_spiel_gin_rummy",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_gin_rummy")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-
-    def test_gin_rummy_observation_is_json(self):
-        env = make(
-            "open_spiel_gin_rummy",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # After dealing, it is player 0's turn (FirstUpcard phase).
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs_p0["phase"], "FirstUpcard")
-        self.assertEqual(obs_p0["current_player"], 0)
-        self.assertFalse(obs_p0["is_terminal"])
-        self.assertEqual(obs_p0["knock_card"], 10)
-        self.assertEqual(obs_p0["stock_size"], 31)
-        self.assertIsNotNone(obs_p0["upcard"])
-        # Player 0 sees their own 10-card hand; opponent's hand is hidden.
-        self.assertEqual(len(obs_p0["hands"]["0"]), 10)
-        self.assertEqual(obs_p0["hands"]["1"], [])
-        self.assertIsNotNone(obs_p0["deadwood"]["0"])
-        self.assertIsNone(obs_p0["deadwood"]["1"])
-
-    def test_gin_rummy_observation_hides_opponent(self):
-        env = make("open_spiel_gin_rummy", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # Each player should only see their own hand in their observation.
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        obs_p1 = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(len(obs_p0["hands"]["0"]), 10)
-        self.assertEqual(obs_p0["hands"]["1"], [])
-        self.assertEqual(obs_p1["hands"]["0"], [])
-        self.assertEqual(len(obs_p1["hands"]["1"]), 10)
-
-    def test_gin_rummy_invalid_action(self):
-        env = make("open_spiel_gin_rummy", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # In FirstUpcard phase only actions 52 (Draw upcard) and 54 (Pass)
-        # are legal, so 0 (the As card) is an invalid action.
-        env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_y_agent_playthrough(self):
-        env = make(
-            "open_spiel_y",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_y")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        rewards = playthrough["rewards"]
-        self.assertEqual(sorted(rewards), [-1.0, 1.0])
-
-    def test_y_manual_playthrough(self):
-        env = make(
-            "open_spiel_y",
-            configuration={"openSpielGameParameters": {"board_size": 8}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # P0 connects the left column (a1..a8), touching all three sides.
-        # Action encoding: row * 8 + col, with rows/cols 0-indexed.
-        moves = [0, 1, 8, 2, 16, 3, 24, 4, 32, 5, 40, 6, 48, 7, 56]
-        for i, action in enumerate(moves):
-            if i % 2 == 0:
-                env.step([{"submission": action}, {"submission": -1}])
-            else:
-                env.step([{"submission": -1}, {"submission": action}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [1, -1])
-        final_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertEqual(final_obs["winner"], "x")
-        self.assertEqual(final_obs["board_size"], 8)
-        self.assertEqual(final_obs["last_move"], "a8")
-        self.assertEqual(final_obs["board"][0][0], "x")
-        self.assertEqual(final_obs["board"][7][0], "x")
-
-    def test_y_invalid_action(self):
-        env = make("open_spiel_y", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_havannah_agent_playthrough(self):
-        env = make(
-            "open_spiel_havannah",
-            configuration={
-                "includeLegalActions": True,
-                "openSpielGameParameters": {"board_size": 4},
-            },
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_havannah")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        rewards = playthrough["rewards"]
-        self.assertIn(sorted(rewards), ([-1.0, 1.0], [0.0, 0.0]))
-
-    def test_havannah_manual_playthrough(self):
-        env = make(
-            "open_spiel_havannah",
-            configuration={"openSpielGameParameters": {"board_size": 4}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # P0 fills the top edge a1..d1 to form a bridge between corners a1 and d1.
-        # Action encoding: x + y * diameter (diameter = 2*board_size - 1 = 7).
-        # P1 plays harmless cells in row 2 (a2, b2, c2 = actions 7, 8, 9).
-        moves = [(0, 0), (1, 7), (0, 1), (1, 8), (0, 2), (1, 9), (0, 3)]
-        for player, action in moves:
-            if player == 0:
-                env.step([{"submission": action}, {"submission": -1}])
-            else:
-                env.step([{"submission": -1}, {"submission": action}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [1, -1])
-        final_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertEqual(final_obs["winner"], "x")
-        self.assertEqual(final_obs["board_size"], 4)
-        self.assertEqual(final_obs["last_move"], "d1")
-        # Top row (y=0) has 4 cells, all filled by x.
-        self.assertEqual(final_obs["board"][0], ["x", "x", "x", "x"])
-        # Row y=1 has 5 cells; first three are o, rest empty.
-        self.assertEqual(final_obs["board"][1][:3], ["o", "o", "o"])
-
-    def test_havannah_invalid_action(self):
-        env = make(
-            "open_spiel_havannah",
-            configuration={"openSpielGameParameters": {"board_size": 4}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # Action 6 falls inside a cut-off corner; it is not a legal cell.
-        env.step([{"submission": 6}, {"submission": -1}])
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_clobber_agent_playthrough(self):
-        env = make(
-            "open_spiel_clobber",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_clobber")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        # Clobber is zero-sum with no draws.
-        rewards = playthrough["rewards"]
-        self.assertEqual(sorted(rewards), [-1.0, 1.0])
-        final_obs = json.loads(playthrough["steps"][-1][0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertIn(final_obs["winner"], ("o", "x"))
-        self.assertEqual(final_obs["rows"], 5)
-        self.assertEqual(final_obs["columns"], 6)
-
-    def test_clobber_manual_playthrough(self):
-        # 2x2 game ends in three moves: a1b1, a2b2, b1b2 -> P0 wins.
-        env = make(
-            "open_spiel_clobber",
-            configuration={"openSpielGameParameters": {"rows": 2, "columns": 2}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 9}, {"submission": -1}])  # P0: a1b1
-        env.step([{"submission": -1}, {"submission": 1}])  # P1: a2b2
-        env.step([{"submission": 12}, {"submission": -1}])  # P0: b1b2
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [1, -1])
-        final_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertEqual(final_obs["winner"], "o")
-        self.assertEqual(final_obs["last_move"], "b1b2")
-        # Only one white piece left at top-right; everything else empty.
-        self.assertEqual(final_obs["board"], [[".", "o"], [".", "."]])
-
-    def test_clobber_invalid_action(self):
-        env = make("open_spiel_clobber", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_checkers_agent_playthrough(self):
-        env = make(
-            "open_spiel_checkers",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_checkers")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        final_obs = json.loads(playthrough["steps"][-1][0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertIn(final_obs["winner"], ("o", "+", "draw"))
-
-    def test_checkers_manual_playthrough(self):
-        env = make("open_spiel_checkers", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        initial_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertFalse(initial_obs["is_terminal"])
-        self.assertIsNone(initial_obs["winner"])
-        self.assertEqual(initial_obs["current_player"], "o")
-        self.assertIsNone(initial_obs["last_move"])
-        self.assertEqual(initial_obs["piece_counts"], {"o": 12, "+": 12, "O": 0, "*": 0})
-        # board[0] is rank 1 (bottom): player 0 pieces on odd columns.
-        self.assertEqual(initial_obs["board"][0], ["o", ".", "o", ".", "o", ".", "o", "."])
-        self.assertEqual(initial_obs["board"][1], [".", "o", ".", "o", ".", "o", ".", "o"])
-        # board[7] is rank 8 (top): player 1 pieces.
-        self.assertEqual(initial_obs["board"][7], [".", "+", ".", "+", ".", "+", ".", "+"])
-        # Action 322 = a3b4 (player 0 moves piece diagonally forward).
-        env.step([{"submission": 322}, {"submission": -1}])
-        after_obs = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(after_obs["current_player"], "+")
-        self.assertEqual(after_obs["last_move"], "a3b4")
-        self.assertEqual(after_obs["move_number"], 1)
-        # a3 is now empty, b4 has player 0 piece.
-        self.assertEqual(after_obs["board"][2][0], ".")  # a3 empty
-        self.assertEqual(after_obs["board"][3][1], "o")  # b4 has piece
-
-    def test_checkers_invalid_action(self):
-        env = make("open_spiel_checkers", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_lines_of_action_agent_playthrough(self):
-        env = make(
-            "open_spiel_lines_of_action",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_lines_of_action")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-
-    def test_lines_of_action_manual_playthrough(self):
-        env = make("open_spiel_lines_of_action", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # Initial board has black (x) on top/bottom rows except corners,
-        # white (o) on left/right columns except corners. Black moves first.
-        initial_obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertFalse(initial_obs["is_terminal"])
-        self.assertIsNone(initial_obs["winner"])
-        self.assertEqual(initial_obs["current_player"], "x")
-        self.assertIsNone(initial_obs["last_move"])
-        # board[0] is rank 1 (bottom). Bottom rank: ".xxxxxx."; rank 2: "o......o".
-        self.assertEqual(initial_obs["board"][0], [".", "x", "x", "x", "x", "x", "x", "."])
-        self.assertEqual(initial_obs["board"][1], ["o", ".", ".", ".", ".", ".", ".", "o"])
-        # Action 142 is "b1-h1" (move b1 piece along the bottom row to h1).
-        env.step([{"submission": 142}, {"submission": -1}])
-        after_obs = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(after_obs["current_player"], "o")
-        self.assertEqual(after_obs["last_move"], "b1-h1")
-        self.assertEqual(after_obs["move_number"], 1)
-        self.assertEqual(after_obs["board"][0], [".", ".", "x", "x", "x", "x", "x", "x"])
-
-    def test_lines_of_action_invalid_action(self):
-        env = make("open_spiel_lines_of_action", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_mancala_agent_playthrough(self):
-        env = make(
-            "open_spiel_mancala",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_mancala")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        final_obs = json.loads(playthrough["steps"][-1][0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertIn(final_obs["winner"], (0, 1, "draw"))
-        self.assertEqual(len(final_obs["board"]), 14)
-        self.assertEqual(final_obs["scores"], [final_obs["stores"]["0"], final_obs["stores"]["1"]])
-
-    def test_mancala_initial_state(self):
-        env = make("open_spiel_mancala", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        # Mancala starts with 4 stones in each of the 12 pits, stores empty.
-        self.assertEqual(obs["board"], [0, 4, 4, 4, 4, 4, 4, 0, 4, 4, 4, 4, 4, 4])
-        self.assertEqual(obs["current_player"], 0)
-        self.assertEqual(obs["scores"], [0, 0])
-        self.assertFalse(obs["is_terminal"])
-        self.assertIsNone(obs["winner"])
-        self.assertIsNone(obs["last_action"])
-
-    def test_mancala_invalid_action(self):
-        env = make("open_spiel_mancala", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_tic_tac_toe_agent_playthrough(self):
-        env = make("open_spiel_tic_tac_toe", debug=True)
-        env.run(["random", "random"])
-        json = env.toJSON()
-        self.assertEqual(json["name"], "open_spiel_tic_tac_toe")
-        self.assertTrue(all([status == "DONE" for status in json["statuses"]]))
-
-    def test_tic_tac_toe_manual_playthrough(self):
-        env = make("open_spiel_tic_tac_toe", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        env.step([{"submission": 0}, {"submission": -1}])
-        env.step([{"submission": -1}, {"submission": 1}])
-        env.step([{"submission": 3}, {"submission": -1}])
-        env.step([{"submission": -1}, {"submission": 4}])
-        env.step([{"submission": 6}, {"submission": -1}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [1, -1])
-
     def test_invalid_action(self):
         env = make("open_spiel_tic_tac_toe", debug=True)
         env.reset()
@@ -641,210 +100,6 @@ class OpenSpielEnvTest(absltest.TestCase):
                 -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
             ],
         )
-
-    def test_amazons_agent_playthrough(self):
-        env = make("open_spiel_amazons", debug=True)
-        env.run(["random", "random"])
-        json_data = env.toJSON()
-        self.assertEqual(json_data["name"], "open_spiel_amazons")
-        self.assertTrue(all(status == "DONE" for status in json_data["statuses"]))
-
-    def test_amazons_manual_playthrough(self):
-        # Walks player 0 (X) through one full Amazons turn (from -> to -> shoot)
-        # by picking the first legal action at each sub-action, and verifies
-        # the proxy advances `phase` through all three values and then resets
-        # to "from" for player O. Picking from `legalActions` keeps this test
-        # robust to pyspiel version differences in the starting layout.
-        env = make("open_spiel_amazons", {"includeLegalActions": True}, debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        for expected_next_phase in ("to", "shoot", "from"):
-            legal = env.state[0]["observation"]["legalActions"]
-            env.step([{"submission": legal[0]}, {"submission": -1}])
-            obs = json.loads(env.state[0]["observation"]["observationString"])
-            if expected_next_phase == "from":
-                # After three sub-actions it is player 1's (O) turn.
-                obs = json.loads(env.state[1]["observation"]["observationString"])
-                self.assertEqual(env.state[0]["status"], "INACTIVE")
-                self.assertEqual(env.state[1]["status"], "ACTIVE")
-                self.assertEqual(obs["current_player"], "o")
-            self.assertEqual(obs["phase"], expected_next_phase)
-            self.assertFalse(obs["is_terminal"])
-        # An arrow (burned square) was fired somewhere on the board.
-        self.assertTrue(any("#" in row for row in obs["board"]))
-
-    def test_amazons_terminal_state(self):
-        # Drive an Amazons game to natural termination using only legal random
-        # actions, then verify the proxy reports a winner and clears the phase.
-        rng = random.Random(0)
-        game = pyspiel.load_game("amazons_proxy")
-        state = game.new_initial_state()
-        while not state.is_terminal():
-            state.apply_action(rng.choice(state.legal_actions()))
-        obs = json.loads(state.observation_string(0))
-        self.assertTrue(obs["is_terminal"])
-        self.assertIn(obs["winner"], ("x", "o"))
-        self.assertIsNone(obs["phase"])
-
-    def test_amazons_invalid_action(self):
-        env = make("open_spiel_amazons", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        json_data = env.toJSON()
-        self.assertEqual(json_data["rewards"][0], open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-        self.assertEqual(json_data["rewards"][1], -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-
-    def test_backgammon_agent_playthrough(self):
-        env = make(
-            "open_spiel_backgammon",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_backgammon")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        final_obs = json.loads(playthrough["steps"][-1][0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-        self.assertIn(final_obs["winner"], ("x", "o", "draw"))
-
-    def test_backgammon_observation_is_json(self):
-        env = make(
-            "open_spiel_backgammon",
-            configuration={"seed": 42, "includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertFalse(obs["is_terminal"])
-        self.assertIsNone(obs["winner"])
-        self.assertIn(obs["current_player"], ("x", "o"))
-        # Seed 42: O plays first with dice [3, 6].
-        self.assertEqual(obs["current_player"], "o")
-        self.assertEqual(obs["dice"], [{"value": 3, "used": False}, {"value": 6, "used": False}])
-        # Standard backgammon starting position (OpenSpiel coords).
-        self.assertEqual(len(obs["board"]), 24)
-        self.assertEqual(obs["board"][0], {"player": "x", "count": 2})
-        self.assertEqual(obs["board"][5], {"player": "o", "count": 5})
-        self.assertEqual(obs["board"][7], {"player": "o", "count": 3})
-        self.assertEqual(obs["board"][11], {"player": "x", "count": 5})
-        self.assertEqual(obs["board"][12], {"player": "o", "count": 5})
-        self.assertEqual(obs["board"][16], {"player": "x", "count": 3})
-        self.assertEqual(obs["board"][18], {"player": "x", "count": 5})
-        self.assertEqual(obs["board"][23], {"player": "o", "count": 2})
-        self.assertEqual(obs["bar"], {"x": 0, "o": 0})
-        self.assertEqual(obs["off"], {"x": 0, "o": 0})
-        # 15 checkers per player on the board.
-        x_count = sum(p["count"] for p in obs["board"] if p and p["player"] == "x")
-        o_count = sum(p["count"] for p in obs["board"] if p and p["player"] == "o")
-        self.assertEqual(x_count, 15)
-        self.assertEqual(o_count, 15)
-
-    def test_backgammon_invalid_action(self):
-        # Seed 7 makes player X (0) move first so the invalid submission
-        # from player 0 below is the one actually evaluated.
-        env = make("open_spiel_backgammon", configuration={"seed": 7}, debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 9999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_dots_and_boxes_agent_playthrough(self):
-        env = make("open_spiel_dots_and_boxes", debug=True)
-        env.run(["random", "random"])
-        json_data = env.toJSON()
-        self.assertEqual(json_data["name"], "open_spiel_dots_and_boxes")
-        self.assertTrue(all(status == "DONE" for status in json_data["statuses"]))
-
-    def test_dots_and_boxes_manual_playthrough(self):
-        # Walks a 2x2 board to a state where P1 closes both top-row boxes back
-        # to back. Confirms the proxy honors the "extra turn" rule (P1 keeps
-        # the move after closing a box), the boxes/scores grids match what we
-        # would expect by hand, and the game eventually terminates with a
-        # winner and all four boxes claimed.
-        env = make(
-            "open_spiel_dots_and_boxes",
-            {
-                "includeLegalActions": True,
-                "openSpielGameParameters": {"num_rows": 2, "num_cols": 2},
-            },
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-
-        def submit(player: int, action: int) -> None:
-            actions = [{"submission": -1}, {"submission": -1}]
-            actions[player] = {"submission": action}
-            env.step(actions)
-
-        # Action layout for num_rows=num_cols=2: horizontal lines 0..5
-        # (h(r,c) = r*2 + c) and vertical lines 6..11 (v(r,c) = 6 + r*3 + c).
-        submit(0, 0)  # P1: h(0,0) — top of box (0,0)
-        submit(1, 1)  # P2: h(0,1) — top of box (0,1)
-        submit(0, 6)  # P1: v(0,0) — left of box (0,0)
-        submit(1, 8)  # P2: v(0,2) — right of box (0,1)
-        submit(0, 7)  # P1: v(0,1) — shared edge between (0,0) and (0,1)
-        submit(1, 4)  # P2: h(2,0) — bottom row, no box closes
-        submit(0, 2)  # P1: h(1,0) closes box (0,0) → score [1,0], P1 keeps turn
-        submit(0, 3)  # P1: h(1,1) closes box (0,1) → score [2,0], P1 keeps turn
-
-        obs0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs0["scores"], [2, 0])
-        self.assertEqual(obs0["boxes"][0], [1, 1])
-        self.assertEqual(obs0["current_player"], "1")
-        self.assertFalse(obs0["is_terminal"])
-        self.assertEqual(obs0["last_action"], {"orientation": "h", "row": 1, "col": 1, "player": "1"})
-
-        # Drain the remaining lines with random legal play. The exact score
-        # split depends on the RNG, but the game must finish with all four
-        # boxes claimed.
-        rng = random.Random(0)
-        while not env.done:
-            current = env.state[0]["observation"]["currentPlayer"]
-            legal = env.state[current]["observation"]["legalActions"]
-            if not legal:
-                break
-            submit(current, rng.choice(legal))
-        obs_final = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(obs_final["is_terminal"])
-        self.assertIn(obs_final["winner"], ("1", "2", "draw"))
-        self.assertEqual(sum(obs_final["scores"]), 4)
-
-    def test_dots_and_boxes_terminal_state(self):
-        # Drive a fresh dots_and_boxes game to termination using only legal
-        # random moves and verify the proxy reports a coherent final state.
-        rng = random.Random(0)
-        game = pyspiel.load_game("dots_and_boxes_proxy")
-        state = game.new_initial_state()
-        while not state.is_terminal():
-            state.apply_action(rng.choice(state.legal_actions()))
-        obs = json.loads(state.observation_string(0))
-        self.assertTrue(obs["is_terminal"])
-        self.assertEqual(obs["current_player"], "")
-        self.assertIn(obs["winner"], ("1", "2", "draw"))
-        self.assertEqual(sum(obs["scores"]), obs["num_rows"] * obs["num_cols"])
-
-    def test_dots_and_boxes_invalid_action(self):
-        env = make("open_spiel_dots_and_boxes", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        json_data = env.toJSON()
-        self.assertEqual(json_data["rewards"][0], open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-        self.assertEqual(json_data["rewards"][1], -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
 
     def test_serialized_game_and_state(self):
         env = make("open_spiel_tic_tac_toe", debug=True)
@@ -907,261 +162,6 @@ class OpenSpielEnvTest(absltest.TestCase):
         self.assertTrue(env.done)
         json_playthrough = env.toJSON()
         self.assertEqual(json_playthrough["rewards"], [-1, 1])
-
-    def test_chess_openings_manually_configured(self):
-        open_spiel_env._register_game_envs(["chess"])
-        openings_path = pathlib.Path(
-            open_spiel_env.GAMES_DIR,
-            "chess/openings.jsonl",
-        )
-        self.assertTrue(openings_path.is_file())
-        with open(openings_path, "r", encoding="utf-8") as f:
-            for line in f:
-                opening = json.loads(line)
-                config = {
-                    "initialActions": opening.pop("initialActions"),
-                    "metadata": opening,
-                }
-                env = make(
-                    "open_spiel_chess",
-                    config,
-                    debug=True,
-                )
-                env.reset()
-                # Setup step
-                env.step(
-                    [
-                        {"submission": pyspiel.INVALID_ACTION},
-                        {"submission": pyspiel.INVALID_ACTION},
-                    ]
-                )
-                obs = env.state[0]["observation"]
-                _, state = pyspiel.deserialize_game_and_state(obs["serializedGameAndState"])
-                self.assertEqual(str(state), opening["fen"])
-                self.assertEqual(str(state), env.toJSON()["configuration"]["metadata"]["fen"])
-
-    def test_chess_openings_configured_with_seed(self):
-        open_spiel_env._register_game_envs(["chess"])
-        config = {
-            "useImage": True,
-            "seed": 1,
-        }
-        env = make(
-            "open_spiel_chess",
-            config,
-            debug=True,
-        )
-        env.reset()
-        # Image config is loaded during setup step.
-        self.assertFalse("imageConfig" in env.configuration)
-        # Setup step
-        env.step(
-            [
-                {"submission": pyspiel.INVALID_ACTION},
-                {"submission": pyspiel.INVALID_ACTION},
-            ]
-        )
-        self.assertTrue("imageConfig" in env.configuration)
-        self.assertEqual(env.configuration["imageConfig"]["color"], "blue")
-        self.assertEqual(env.configuration["imageConfig"]["pieceSet"], "cardinal")
-        self.assertTrue("imageConfig" in env.state[0]["observation"])
-
-    def test_default_repeated_poker(self):
-        env = make("open_spiel_repeated_poker", {"openSpielGameString": TEST_REPEATED_POKER_GAME_STRING})
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        # Default repeated_poker now includes hand odds calculations which take
-        # ~1-2s per action, so we avoid stepping until the end of the game.
-        for i in range(2):
-            if i % 2 == 0:
-                env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                env.step([{"submission": 0}, {"submission": -1}])
-        self.assertEqual(len(env.os_state.acpc_hand_histories()), 2)
-        state_dict = json.loads(str(env.os_state))
-        self.assertTrue("current_universal_poker_json" in state_dict)
-        current_hand_dict = json.loads(state_dict["current_universal_poker_json"])
-        self.assertTrue("odds" in current_hand_dict)
-        self.assertEqual(len(current_hand_dict["odds"]), 4)
-
-    def test_repeated_pokerkit(self):
-        pokerkit_game_str = (
-            "python_repeated_pokerkit("
-            "bet_size_schedule=,"
-            "blind_schedule=,"
-            "bring_in_schedule=,"
-            "first_button_player=-1,"
-            "max_num_hands=20,"
-            "pokerkit_game_params=python_pokerkit_wrapper("
-            "blinds=1 2,"
-            "num_players=2,"
-            "stack_sizes=200 200,"
-            "variant=NoLimitTexasHoldem),"
-            "reset_stacks=True,"
-            "rotate_dealer=True)"
-        )
-        envs = open_spiel_env._register_game_envs([pokerkit_game_str])
-        env = make(envs["open_spiel_python_repeated_pokerkit"])
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        for i in range(20):
-            if i % 2 == 0:
-                env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [0.0, 0.0])
-
-    def test_default_repeated_pokerkit_loads(self):
-        env = make("open_spiel_python_repeated_pokerkit")
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        for i in range(100):
-            if i % 2 == 0:
-                env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [0.0, 0.0])
-
-    def test_poker_set_num_hands(self):
-        num_hands = 2
-        config = {
-            "openSpielGameString": TEST_REPEATED_POKER_GAME_STRING,
-            "setNumHands": num_hands,
-        }
-        env = make(
-            "open_spiel_repeated_poker",
-            config,
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Initial setup step.
-        for i in range(num_hands):
-            if i % 2 == 0:
-                env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(env.done)
-        self.assertEqual(env.toJSON()["rewards"], [0.0, 0.0])
-
-    def test_repeated_poker_preset_hands_replay(self):
-        base_env = make(
-            "open_spiel_repeated_poker",
-            {"openSpielGameString": TEST_REPEATED_POKER_GAME_STRING, "setNumHands": 2},
-            debug=True,
-        )
-        base_env.reset()
-        base_env.step([{"submission": -1}, {"submission": -1}])
-        for i in range(2):
-            if i % 2 == 0:
-                base_env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                base_env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(base_env.done)
-        base_history = [int(action) for action in base_env.info["actionHistory"]]
-        replay_state = base_env.os_game.new_initial_state()
-        preset_hands: list[list[int]] = []
-        for action in base_history:
-            if replay_state.is_chance_node():
-                hand_idx = len(replay_state.acpc_hand_histories())
-                while len(preset_hands) <= hand_idx:
-                    preset_hands.append([])
-                preset_hands[hand_idx].append(action)
-            replay_state.apply_action(action)
-        if preset_hands and preset_hands[0]:
-            preset_hands[0].append(preset_hands[0][-1])
-        preset_env = make(
-            "open_spiel_repeated_poker",
-            {
-                "openSpielGameString": TEST_REPEATED_POKER_GAME_STRING,
-                "setNumHands": 2,
-                "presetHands": [hand[:] for hand in preset_hands],
-            },
-            debug=True,
-        )
-        preset_env.reset()
-        preset_env.step([{"submission": -1}, {"submission": -1}])
-        for i in range(2):
-            if i % 2 == 0:
-                preset_env.step([{"submission": -1}, {"submission": 0}])
-            else:
-                preset_env.step([{"submission": 0}, {"submission": -1}])
-        self.assertTrue(preset_env.done)
-        self.assertEqual(
-            [int(action) for action in preset_env.info["actionHistory"]],
-            base_history,
-        )
-        self.assertEqual(
-            preset_env.info["presetHands"],
-            [hand[:] for hand in preset_hands],
-        )
-        self.assertLess(
-            preset_env.info["presetHandsState"]["next_index"][0],
-            len(preset_env.info["presetHands"][0]),
-        )
-
-    def test_repeated_poker_preset_hands_runs_out(self):
-        env = make(
-            "open_spiel_repeated_poker",
-            {"openSpielGameString": TEST_REPEATED_POKER_GAME_STRING, "presetHands": [[0]]},
-            debug=True,
-        )
-        env.reset()
-        with self.assertRaisesRegex(ValueError, "presetHands"):
-            env.step([{"submission": -1}, {"submission": -1}])
-
-    def test_repeated_poker_preset_hands_conflicts_with_use_openings(self):
-        env = make(
-            "open_spiel_repeated_poker",
-            {
-                "openSpielGameString": TEST_REPEATED_POKER_GAME_STRING,
-                "presetHands": [[0, 1, 2, 3, 4, 5, 6, 7, 8]],
-                "useOpenings": True,
-            },
-            debug=True,
-        )
-        env.reset()
-        with self.assertRaisesRegex(ValueError, "useOpenings"):
-            env.step([{"submission": -1}, {"submission": -1}])
-
-    def test_repeated_poker_load_preset_hands_loads_file(self):
-        env = make(
-            "open_spiel_repeated_poker",
-            {"openSpielGameString": TEST_REPEATED_POKER_GAME_STRING, "loadPresetHands": True, "seed": 0},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])
-        self.assertIn("presetHands", env.configuration)
-        preset_hands = env.configuration["presetHands"]
-        self.assertEqual(len(preset_hands), 100)
-        self.assertTrue(all(len(hand) == 9 for hand in preset_hands))
-
-    def test_repeated_poker_load_preset_hands_requires_seed(self):
-        env = make(
-            "open_spiel_repeated_poker",
-            {"openSpielGameString": TEST_REPEATED_POKER_GAME_STRING, "loadPresetHands": True},
-            debug=True,
-        )
-        env.reset()
-        with self.assertRaisesRegex(ValueError, "seed"):
-            env.step([{"submission": -1}, {"submission": -1}])
-
-    def test_repeated_poker_load_preset_hands_conflicts_with_manual(self):
-        env = make(
-            "open_spiel_repeated_poker",
-            {
-                "openSpielGameString": TEST_REPEATED_POKER_GAME_STRING,
-                "loadPresetHands": True,
-                "seed": 0,
-                "presetHands": [[0, 1, 2, 3, 4, 5, 6, 7, 8]],
-            },
-            debug=True,
-        )
-        env.reset()
-        with self.assertRaisesRegex(ValueError, "loadPresetHands"):
-            env.step([{"submission": -1}, {"submission": -1}])
 
     def test_game_params_override(self):
         """Test that openSpielGameParameters can override default params."""
@@ -1263,38 +263,7 @@ class OpenSpielEnvTest(absltest.TestCase):
         self.assertIsNotNone(obs.legalActionStrings)
         self.assertEqual(len(obs.legalActions), 9)  # All 9 squares available
 
-    # --- Simultaneous move game tests ---
-
-    def test_goofspiel_agent_playthrough(self):
-        open_spiel_env._register_game_envs(["goofspiel(num_cards=4,points_order=descending,returns_type=total_points)"])
-        env = make("open_spiel_goofspiel", debug=True)
-        env.run(["random", "random"])
-        json = env.toJSON()
-        self.assertEqual(json["name"], "open_spiel_goofspiel")
-        self.assertTrue(all([status == "DONE" for status in json["statuses"]]))
-        # Both players should have rewards (total_points mode).
-        self.assertEqual(len(json["rewards"]), 2)
-        self.assertTrue(all(r is not None for r in json["rewards"]))
-
-    def test_goofspiel_manual_playthrough(self):
-        open_spiel_env._register_game_envs(["goofspiel(num_cards=4,points_order=descending,returns_type=total_points)"])
-        env = make("open_spiel_goofspiel", debug=True)
-        env.reset()
-        # Initial setup step.
-        env.step([{"submission": -1}, {"submission": -1}])
-        # After setup, both players should be ACTIVE (simultaneous node).
-        self.assertEqual(env.state[0]["status"], "ACTIVE")
-        self.assertEqual(env.state[1]["status"], "ACTIVE")
-        # Play all 4 rounds: both players submit actions each step.
-        # With descending point order and 4 cards, there are 4 bidding rounds.
-        # Legal actions are card indices (0-3 initially).
-        for _ in range(4):
-            if env.done:
-                break
-            env.step([{"submission": 0}, {"submission": 0}])
-        self.assertTrue(env.done)
-        json = env.toJSON()
-        self.assertTrue(all([status == "DONE" for status in json["statuses"]]))
+    # --- Simultaneous move game dispatch ---
 
     def test_simultaneous_invalid_action(self):
         open_spiel_env._register_game_envs(["goofspiel(num_cards=4,points_order=descending,returns_type=total_points)"])
@@ -1309,211 +278,6 @@ class OpenSpielEnvTest(absltest.TestCase):
         # Player 0 submitted invalid, so gets the penalty.
         self.assertEqual(json["rewards"][0], open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
         self.assertEqual(json["rewards"][1], -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-
-    def test_matching_pennies_manual(self):
-        open_spiel_env._register_game_envs(["matching_pennies_3p"])
-        env = make("open_spiel_matching_pennies_3p", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}, {"submission": -1}])  # Setup.
-        # All 3 players should be ACTIVE.
-        for pid in range(3):
-            self.assertEqual(env.state[pid]["status"], "ACTIVE")
-        # All play Heads (action 0).
-        env.step([{"submission": 0}, {"submission": 0}, {"submission": 0}])
-        self.assertTrue(env.done)
-        json = env.toJSON()
-        self.assertTrue(all([status == "DONE" for status in json["statuses"]]))
-        self.assertEqual(len(json["rewards"]), 3)
-
-    def test_repeated_prisoners_dilemma_agent_playthrough(self):
-        """Test repeated Prisoner's Dilemma with random agents for 10 rounds."""
-        env = make(
-            "open_spiel_repeated_game",
-            {"openSpielGameParameters": {"num_repetitions": 10}},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        json = env.toJSON()
-        self.assertEqual(json["name"], "open_spiel_repeated_game")
-        self.assertTrue(all(s == "DONE" for s in json["statuses"]))
-        self.assertEqual(len(json["rewards"]), 2)
-        self.assertTrue(all(r is not None for r in json["rewards"]))
-
-    def test_repeated_prisoners_dilemma_mutual_cooperate(self):
-        """Both players cooperate every round. Expected reward: 5 * 10 = 50 each."""
-        env = make(
-            "open_spiel_repeated_game",
-            {"openSpielGameParameters": {"num_repetitions": 10}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        for _ in range(10):
-            self.assertFalse(env.done)
-            self.assertEqual(env.state[0]["status"], "ACTIVE")
-            self.assertEqual(env.state[1]["status"], "ACTIVE")
-            env.step([{"submission": 0}, {"submission": 0}])  # Both cooperate.
-        self.assertTrue(env.done)
-        json = env.toJSON()
-        self.assertEqual(json["rewards"], [50.0, 50.0])
-
-    def test_repeated_prisoners_dilemma_mutual_defect(self):
-        """Both players defect every round. Expected reward: 1 * 10 = 10 each."""
-        env = make(
-            "open_spiel_repeated_game",
-            {"openSpielGameParameters": {"num_repetitions": 10}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        for _ in range(10):
-            env.step([{"submission": 1}, {"submission": 1}])  # Both defect.
-        self.assertTrue(env.done)
-        json = env.toJSON()
-        self.assertEqual(json["rewards"], [10.0, 10.0])
-
-    def test_repeated_prisoners_dilemma_asymmetric(self):
-        """P0 always cooperates, P1 always defects. P0 gets 0, P1 gets 100."""
-        env = make(
-            "open_spiel_repeated_game",
-            {"openSpielGameParameters": {"num_repetitions": 10}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        for _ in range(10):
-            env.step([{"submission": 0}, {"submission": 1}])  # P0 cooperate, P1 defect.
-        self.assertTrue(env.done)
-        json = env.toJSON()
-        self.assertEqual(json["rewards"], [0.0, 100.0])
-
-    def test_oshi_zumo_agent_playthrough(self):
-        env = make(
-            "open_spiel_oshi_zumo",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_oshi_zumo")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        rewards = playthrough["rewards"]
-        self.assertEqual(len(rewards), 2)
-        # Zero-sum game.
-        self.assertAlmostEqual(rewards[0] + rewards[1], 0.0)
-        # Verify the proxy returns JSON observations.
-        obs_str = playthrough["steps"][1][0]["observation"]["observationString"]
-        obs = json.loads(obs_str)
-        self.assertIn("field", obs)
-        self.assertIn("coins", obs)
-        self.assertIn("wrestler_position", obs)
-
-    def test_oshi_zumo_manual_playthrough(self):
-        env = make(
-            "open_spiel_oshi_zumo",
-            {"openSpielGameParameters": {"coins": 4, "size": 1, "horizon": 100}},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        # Both players bid all 4 coins -> wrestler stays at center, coins exhausted.
-        env.step([{"submission": 4}, {"submission": 4}])
-        self.assertTrue(env.done)
-        json_out = env.toJSON()
-        self.assertTrue(all(status == "DONE" for status in json_out["statuses"]))
-        # Wrestler still at center on a size=1 field -> draw.
-        self.assertEqual(json_out["rewards"], [0.0, 0.0])
-        obs = json.loads(json_out["steps"][-1][0]["observation"]["observationString"])
-        self.assertEqual(obs["coins"], [0, 0])
-        self.assertTrue(obs["is_terminal"])
-
-    def test_oshi_zumo_invalid_action(self):
-        env = make("open_spiel_oshi_zumo", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        # Bidding more coins than the player has is illegal.
-        env.step([{"submission": 999}, {"submission": 0}])
-        self.assertTrue(env.done)
-        json_out = env.toJSON()
-        self.assertEqual(json_out["rewards"][0], open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-        self.assertEqual(json_out["rewards"][1], -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD)
-
-    def test_snake_agent_playthrough(self):
-        env = make(
-            "open_spiel_snake",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        json_out = env.toJSON()
-        self.assertEqual(json_out["name"], "open_spiel_snake")
-        self.assertTrue(all(status == "DONE" for status in json_out["statuses"]))
-
-    def test_snake_observation_is_json(self):
-        env = make(
-            "open_spiel_snake",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs["num_rows"], 10)
-        self.assertEqual(obs["num_columns"], 10)
-        self.assertEqual(obs["num_players"], 2)
-        self.assertEqual(len(obs["board"]), 10)
-        self.assertTrue(all(len(row) == 10 for row in obs["board"]))
-        self.assertEqual(obs["snakes"][0]["body"], [[1, 1]])
-        self.assertEqual(obs["snakes"][1]["body"], [[8, 8]])
-        self.assertTrue(obs["snakes"][0]["alive"])
-        self.assertTrue(obs["snakes"][1]["alive"])
-        self.assertEqual(obs["scores"], [0.0, 0.0])
-        self.assertEqual(obs["current_player"], 0)
-        self.assertFalse(obs["is_terminal"])
-        self.assertIsNone(obs["winner"])
-        self.assertEqual(obs["turn"], 0)
-        # Food is placed on an empty square.
-        self.assertIsNotNone(obs["food"])
-        fr, fc = obs["food"]
-        self.assertEqual(obs["board"][fr][fc], "*")
-
-    def test_snake_manual_playthrough(self):
-        # Sequential implementation of simultaneous play: each player submits
-        # in turn, then the buffered moves are applied together. Drive both
-        # snakes off the board on the first round so both die and the game
-        # terminates immediately.
-        env = make("open_spiel_snake", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        # P0 at (1,1): UP -> (0,1)? still on board. Send LEFT (2) twice to die.
-        # First round: both submit a move; resolution moves snakes and may kill.
-        # P0 LEFT from (1,1) -> (1,0) (alive). P1 RIGHT from (8,8) -> (8,9) (alive).
-        # Second round: P0 LEFT from (1,0) -> (1,-1) (wall, dies).
-        # P1 RIGHT from (8,9) -> (8,10) (wall, dies).
-        env.step([{"submission": 2}, {"submission": -1}])  # P0 LEFT
-        env.step([{"submission": -1}, {"submission": 3}])  # P1 RIGHT
-        env.step([{"submission": 2}, {"submission": -1}])  # P0 LEFT (off board)
-        env.step([{"submission": -1}, {"submission": 3}])  # P1 RIGHT (off board)
-        self.assertTrue(env.done)
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(obs["is_terminal"])
-        self.assertFalse(obs["snakes"][0]["alive"])
-        self.assertFalse(obs["snakes"][1]["alive"])
-
-    def test_snake_invalid_action(self):
-        env = make("open_spiel_snake", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        json_out = env.toJSON()
-        self.assertEqual(
-            json_out["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
 
     def test_simultaneous_agent_error(self):
         open_spiel_env._register_game_envs(["goofspiel(num_cards=4,points_order=descending,returns_type=total_points)"])
@@ -1530,215 +294,6 @@ class OpenSpielEnvTest(absltest.TestCase):
         json = env.toJSON()
         self.assertEqual(json["rewards"], [None, None])
         self.assertEqual(json["statuses"], ["ERROR", "ERROR"])
-
-    def test_ultimate_tic_tac_toe_agent_playthrough(self):
-        env = make(
-            "open_spiel_ultimate_tic_tac_toe",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_ultimate_tic_tac_toe")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-
-    def test_ultimate_tic_tac_toe_manual_playthrough(self):
-        env = make("open_spiel_ultimate_tic_tac_toe", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step
-
-        # Sequential game: only the current player submits
-        # Step 1: Player 0 selects sub-grid 6
-        env.step([{"submission": 6}, {"submission": -1}])
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs_p0["active_subgrid"], 6)
-        self.assertEqual(obs_p0["phase"], "choose_cell")
-        self.assertEqual(obs_p0["current_player"], "x")
-
-        # Step 2: Player 0 selects cell 1 of sub-grid 6
-        env.step([{"submission": 1}, {"submission": -1}])
-        obs_p1 = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(obs_p1["active_subgrid"], 1)
-        self.assertEqual(obs_p1["phase"], "choose_cell")
-        self.assertEqual(obs_p1["current_player"], "o")
-        self.assertEqual(obs_p1["board"][6][1], "x")
-
-        # Step 3: Player 1 selects cell 2 of sub-grid 1
-        env.step([{"submission": -1}, {"submission": 2}])
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs_p0["active_subgrid"], 2)
-        self.assertEqual(obs_p0["phase"], "choose_cell")
-        self.assertEqual(obs_p0["current_player"], "x")
-        self.assertEqual(obs_p0["board"][1][2], "o")
-
-    def test_ultimate_tic_tac_toe_invalid_action(self):
-        env = make("open_spiel_ultimate_tic_tac_toe", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
-
-    def test_python_ant_foraging_agent_playthrough(self):
-        env = make(
-            "open_spiel_python_ant_foraging",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_python_ant_foraging")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-        # Cooperative game: both ants share the same reward.
-        rewards = playthrough["rewards"]
-        self.assertEqual(rewards[0], rewards[1])
-        final_obs = json.loads(playthrough["steps"][-1][0]["observation"]["observationString"])
-        self.assertTrue(final_obs["is_terminal"])
-
-    def test_python_ant_foraging_observation_is_json(self):
-        env = make(
-            "open_spiel_python_ant_foraging",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs["grid_size"], 8)
-        self.assertEqual(obs["num_ants"], 2)
-        self.assertEqual(obs["num_food"], 3)
-        self.assertEqual(obs["max_turns"], 50)
-        self.assertEqual(obs["turn"], 0)
-        self.assertEqual(obs["food_collected"], 0)
-        self.assertEqual(obs["score"], 0)
-        self.assertEqual(obs["current_player"], 0)
-        self.assertFalse(obs["is_terminal"])
-        # Grid is grid_size x grid_size with the nest marked.
-        self.assertEqual(len(obs["grid"]), 8)
-        self.assertTrue(all(len(row) == 8 for row in obs["grid"]))
-        nest_r, nest_c = obs["nest_position"]
-        self.assertEqual(obs["grid"][nest_r][nest_c], "N")
-        # Both ants start at the nest, carrying nothing.
-        self.assertEqual(obs["ant_positions"], [[nest_r, nest_c], [nest_r, nest_c]])
-        self.assertEqual(obs["carrying_food"], [False, False])
-        # Food cells exist on the grid.
-        self.assertEqual(len(obs["food_positions"]), 3)
-        for fr, fc in obs["food_positions"]:
-            self.assertEqual(obs["grid"][fr][fc], "F")
-        # Pheromone grids are initially zero everywhere.
-        for grid_name in ("pheromone_to_food", "pheromone_to_nest"):
-            grid = obs[grid_name]
-            self.assertEqual(len(grid), 8)
-            self.assertTrue(all(v == 0.0 for row in grid for v in row))
-
-    def test_python_ant_foraging_manual_playthrough(self):
-        env = make(
-            "open_spiel_python_ant_foraging",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        # Sequential game: ants alternate. Move ant 0 first, then ant 1.
-        env.step([{"submission": 0}, {"submission": -1}])  # Ant 0 stays.
-        obs = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(obs["current_player"], 1)
-        self.assertEqual(obs["turn"], 0)  # Turn only advances after both ants move.
-        env.step([{"submission": -1}, {"submission": 0}])  # Ant 1 stays.
-        obs = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertEqual(obs["current_player"], 0)
-        self.assertEqual(obs["turn"], 1)  # Full round completed.
-
-    def test_python_ant_foraging_invalid_action(self):
-        env = make("open_spiel_python_ant_foraging", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"][0],
-            open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-        )
-        self.assertEqual(
-            playthrough["rewards"][1],
-            -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-        )
-
-    def test_negotiation_agent_playthrough(self):
-        env = make(
-            "open_spiel_negotiation",
-            configuration={"includeLegalActions": True},
-            debug=True,
-        )
-        env.run(["random", "random"])
-        playthrough = env.toJSON()
-        self.assertEqual(playthrough["name"], "open_spiel_negotiation")
-        self.assertTrue(all(status == "DONE" for status in playthrough["statuses"]))
-
-    def test_negotiation_observation_is_json_and_hides_opponent_utilities(self):
-        env = make("open_spiel_negotiation", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step (chance).
-        obs_p0 = json.loads(env.state[0]["observation"]["observationString"])
-        obs_p1 = json.loads(env.state[1]["observation"]["observationString"])
-        self.assertEqual(obs_p0["current_player"], 0)
-        self.assertEqual(obs_p0["turn_type"], "proposal")
-        self.assertEqual(len(obs_p0["item_pool"]), 3)
-        self.assertEqual(len(obs_p0["my_utilities"]), 3)
-        self.assertEqual(obs_p0["viewing_player"], 0)
-        self.assertEqual(obs_p1["viewing_player"], 1)
-        # Each player only sees their own utility vector; the other's stays hidden.
-        self.assertNotIn("opponent_utilities", obs_p0)
-        self.assertEqual(obs_p0["params"]["accept_action"], 216)
-        self.assertEqual(obs_p0["params"]["num_distinct_proposals"], 217)
-        self.assertFalse(obs_p0["is_terminal"])
-        self.assertEqual(obs_p0["proposals"], [])
-        self.assertEqual(obs_p0["utterances"], [])
-
-    def test_negotiation_acceptance_terminates_and_assigns_returns(self):
-        env = make("open_spiel_negotiation", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step (chance).
-        # Player 0 proposes [0, 0, 0] (action 0), then utters [0, 0, 0] (action 217).
-        env.step([{"submission": 0}, {"submission": -1}])
-        env.step([{"submission": 217}, {"submission": -1}])
-        # Player 1 accepts the proposal (action = num_distinct_proposals - 1 = 216).
-        env.step([{"submission": -1}, {"submission": 216}])
-        self.assertTrue(env.done)
-        final = json.loads(env.state[0]["observation"]["observationString"])
-        self.assertTrue(final["agreement_reached"])
-        self.assertTrue(final["is_terminal"])
-        # The accepting player keeps every item in the pool, so the proposer
-        # (who offered themselves nothing) ends with zero utility.
-        rewards = env.toJSON()["rewards"]
-        self.assertEqual(rewards[0], 0.0)
-        self.assertGreater(rewards[1], 0.0)
-        self.assertEqual(final["winner"], 1)
-        # The accept action shows up in the proposals history with a flag.
-        self.assertEqual(final["proposals"][-1], {"player": 1, "accept": True})
-
-    def test_negotiation_invalid_action(self):
-        env = make("open_spiel_negotiation", debug=True)
-        env.reset()
-        env.step([{"submission": -1}, {"submission": -1}])  # Setup step.
-        env.step([{"submission": 9999}, {"submission": -1}])  # Invalid action.
-        self.assertTrue(env.done)
-        playthrough = env.toJSON()
-        self.assertEqual(
-            playthrough["rewards"],
-            [
-                open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-                -open_spiel_env.DEFAULT_INVALID_ACTION_REWARD,
-            ],
-        )
 
 
 if __name__ == "__main__":
